requirements.txt
# Basic utilities
requests==2.26.0
pandas==2.0.3

# Data visualization
matplotlib==3.4.3

# Testing
pytest==7.1.3
pytest-cov==4.0.0
nbval==0.9.6 # For testing Jupyter Notebooks

# Code formatting
black==22.3.0

# Code linting
pylint==2.15.3

# Database
pymysql==1.0.2
psycopg2-binary==2.9.1
python-dotenv==0.19.2
mock
